+++ config/settings/base.py
@@ -576,7 +576,7 @@ SITE_INSTANCE_TITLE = env.str('SITE_INSTANCE_TITLE', 'CUBI SODAR')
 
 
 # General API settings
-SODAR_API_DEFAULT_VERSION = '0.11.1'
+SODAR_API_DEFAULT_VERSION = '0.11.2'
 SODAR_API_ALLOWED_VERSIONS = [
     '0.7.0',
     '0.7.1',
@@ -586,6 +586,7 @@ SODAR_API_ALLOWED_VERSIONS = [
     '0.10.1',
     '0.11.0',
     '0.11.1',
+    '0.11.2',
 ]
 SODAR_API_MEDIA_TYPE = 'application/vnd.bihealth.sodar+json'
 SODAR_API_DEFAULT_HOST = env.url(