+++ samplesheets/views_api.py
@@ -206,9 +206,8 @@ class SheetImportAPIView(SheetImportMixin, SODARAPIBaseProjectMixin, APIView):
         old_inv = Investigation.objects.filter(
             project=project, active=True
         ).first()
-        action = 'replace' if old_inv else 'create'
-        zip_file = None
 
+        zip_file = None
         if len(request.FILES) == 0:
             raise ParseError('No files provided')
 
@@ -229,9 +228,8 @@ class SheetImportAPIView(SheetImportMixin, SODARAPIBaseProjectMixin, APIView):
                 raise ParseError('Failed to parse TSV files: {}'.format(ex))
 
         # Handle import
-        tl_event = self.create_timeline_event(
-            project=project, action='replace' if old_inv else 'import'
-        )
+        action = 'replace' if old_inv else 'create'
+        tl_event = self.create_timeline_event(project=project, action=action)
 
         try:
             investigation = sheet_io.import_isa(
@@ -245,13 +243,6 @@ class SheetImportAPIView(SheetImportMixin, SODARAPIBaseProjectMixin, APIView):
         except Exception as ex:
             self.handle_import_exception(ex, tl_event, ui_mode=False)
             raise APIException(str(ex))
-
-        if tl_event:
-            tl_event.add_object(
-                obj=investigation,
-                label='investigation',
-                name=investigation.title,
-            )
 
         # Handle replace
         if old_inv:
@@ -270,6 +261,13 @@ class SheetImportAPIView(SheetImportMixin, SODARAPIBaseProjectMixin, APIView):
                         ex_msg if ex_msg else 'See SODAR error log'
                     )
                 )
+
+        if tl_event:
+            tl_event.add_object(
+                obj=investigation,
+                label='investigation',
+                name=investigation.title,
+            )
 
         # Finalize import
         isa_version = (